// src/app/Models/PostData_API_Class/SaveAndFlowStartClass.ts
import { uploadFileClass } from "../uploadFileClass";

// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface ForgetSaveAndFlowStartClass {
  FlowApp:     SFlowApp;
  FlowDynamic: FlowDynamic;
}

export interface SFlowApp {
  FlowApps: FlowAppElement[];
  EmpID:    string;
  EmpCode:  string;
  EmpNameC: string;
  State:    string;
}

export interface FlowAppElement {
  EmpID:           string;
  EmpCode:         string;
  EmpNameC:        string;
  Date:            string;
  RoteDateTimeB:   string;
  RoteDateTimeE:   string;
  CardDateTimeB:   string;
  CardDateTimeE:   string;
  DateB:           string;
  DateE:           string;
  TimeB:           string;
  TimeE:           string;
  DateTimeB:       string;
  DateTimeE:       string;
  CauseID1:        string;
  CauseName1:      string;
  CauseID2:        string;
  CauseName2:      string;
  Note:            string;
  Info:            string;
  UploadFile:      any[];
  MailBody:        string;
  State:           string;
  ExceptionalCode: string;
  ExceptionalName: string;
}


export interface FlowDynamic {
  FlowNode: string;
  RoleID:   string;
  EmpID:    string;
  DeptID:   string;
  PosID:   string;
}



export class SaveAndFlowStartClass {
  EmpID: string;
  EmpNameC: string;
  Date:string;
  DateB: string;
  DateE: string;
  TimeB: string;
  TimeE: string;
  DateTimeB: string;
  DateTimeE: string;
  CauseID1: string;
  CauseName1: string;
  Note: string;
  ReviewEmpID: string;
  UploadFile: uploadFileClass[]
  ExceptionalCode: string
  ExceptionalName: string
}